fix: perf optimizations for key sequence creation
