+++ packages/marko/src/runtime/components/dom-data.js
@@ -1,18 +1,4 @@
-var counter = 0;
-var seed = "M" + Math.random().toFixed(5);
-var WeakMap =
-  global.WeakMap ||
-  function WeakMap() {
-    var id = seed + counter++;
-    return {
-      get: function(ref) {
-        return ref[id];
-      },
-      set: function(ref, value) {
-        ref[id] = value;
-      }
-    };
-  };
+var WeakMap = require("../helpers/_weak-map");
 
 module.exports = {
   ___vPropsByDOMNode: new WeakMap(),
